+++ src/sll/include/sll/_internal/environment.h
@@ -1,0 +1,10 @@
+#ifndef __SLL__INTERNAL_ENVIRONMENT_H__
+#define __SLL__INTERNAL_ENVIRONMENT_H__ 1
+
+
+
+void _init_environment(void);
+
+
+
+#endif