+++ src/sll/include/sll/environment.h
@@ -1,11 +1,11 @@
-#ifndef __SLL_ENV_H__
-#define __SLL_ENV_H__ 1
+#ifndef __SLL_ENVIRONMENT_H__
+#define __SLL_ENVIRONMENT_H__ 1
 #include <sll/common.h>
 #include <sll/types.h>
 /**
  * \flags group
  * \name Environment Variables
- * \group env
+ * \group environment
  * \desc Docs!
  */
 
@@ -14,7 +14,7 @@
 /**
  * \flags var
  * \name sll_env_path
- * \group env
+ * \group environment
  * \desc Docs!
  * \type const sll_search_path_t*
  */
@@ -23,22 +23,35 @@ __SLL_EXTERNAL extern const sll_search_path_t* sll_env_path;
 
 
 /**
- * \flags check_output func
- * \name sll_get_environment_variable
- * \group env
+ * \flags func
+ * \name sll_expand_environment_variable
+ * \group environment
  * \desc Docs!
  * \arg const sll_string_t* k
  * \arg sll_string_t* o
  * \ret sll_bool_t
  */
-__SLL_EXTERNAL __SLL_CHECK_OUTPUT sll_bool_t sll_get_environment_variable(const sll_string_t* k,sll_string_t* o);
+__SLL_EXTERNAL sll_bool_t sll_expand_environment_variable(const sll_string_t* k,sll_string_t* o);
+
+
+
+/**
+ * \flags func
+ * \name sll_get_environment_variable
+ * \group environment
+ * \desc Docs!
+ * \arg const sll_string_t* k
+ * \arg sll_string_t* o
+ * \ret sll_bool_t
+ */
+__SLL_EXTERNAL sll_bool_t sll_get_environment_variable(const sll_string_t* k,sll_string_t* o);
 
 
 
 /**
  * \flags func
  * \name sll_remove_environment_variable
- * \group env
+ * \group environment
  * \desc Docs!
  * \arg const sll_string_t* k
  */
@@ -49,7 +62,7 @@ __SLL_EXTERNAL void sll_remove_environment_variable(const sll_string_t* k);
 /**
  * \flags func
  * \name sll_set_environment_variable
- * \group env
+ * \group environment
  * \desc Docs!
  * \arg const sll_string_t* k
  * \arg const sll_string_t* v