+++ src/sll/platform/linux/path.c
@@ -4,7 +4,7 @@
 #include <sll/api/path.h>
 #include <sll/common.h>
 #include <sll/data.h>
-#include <sll/env.h>
+#include <sll/environment.h>
 #include <sll/memory.h>
 #include <sll/platform/path.h>
 #include <sll/string.h>